useLocation separated, README update, There's a BUG BUG that I need to solve
