+++ Front-End/src/components/Map.js
@@ -4,7 +4,7 @@ import {
     StyleSheet,
     ActivityIndicator
 } from 'react-native';
-import MapView, {Polyline} from 'react-native-maps';
+import MapView, {Polyline, Circle} from 'react-native-maps';
 
 //Application Imports
 import {Context as LocationContext} from '../context/LocationContext';
@@ -33,7 +33,14 @@ const Map = () => {
             latitudeDelta: 0.01,
             longitudeDelta: 0.01
           }}
-        />
+        >
+            <Circle
+                center={currentLocation.coords}
+                radius={30}
+                strokeColor="rgba(158, 158, 255, 1.0"
+                fillColor="rgba(227, 18, 28, 0.8)"
+            />
+        </MapView>
     );
 };
