+++ Front-End/src/screens/ResolveAuthScreen.js
@@ -1,3 +1,4 @@
+// Imports
 import React, {useEffect, useContext} from 'react';
 import { Context as AuthContext} from '../context/AuthContext';
